requirements.txt
apache-airflow==2.9.0
beautifulsoup4
requests
pandas
